requirements.txt
# Core LLM and AI
anthropic>=0.40.0

# Data Validation
pydantic>=2.0.0

# PDF Processing
PyPDF2>=3.0.0

# Blockchain & Web3
web3>=6.0.0

# API Development
fastapi>=0.109.0
uvicorn>=0.27.0

# CLI Tools
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0